+++ src/ralph_assets/tests/functional/tests_data_importer.py
@@ -103,7 +103,7 @@ class TestDataImporter(object):
         csv_string = '\n'.join([header, rows_string])
         return csv_string
 
-    def _get_csv_data(self, use_existing=True):
+    def _get_csv_data(self, important_data={}, use_existing=True):
         """
         Return dict, where:
             *keys* are names of model fields (model is set by
@@ -129,6 +129,9 @@ class TestDataImporter(object):
                     field_value += '-new'
             csv_data[field.name] = field_value
         obj.delete()
+        if not important_data:
+            important_data = {}
+        csv_data.update(important_data)
         return csv_data
 
     def _check_form_errors(self, response, step):
@@ -202,8 +205,7 @@ class TestDataImporter(object):
         """
         self.excluded_fields.remove('id')
         updated_obj = self.ModelFactory()
-        csv_data = self._get_csv_data()
-        csv_data['id'] = updated_obj.id
+        csv_data = self._get_csv_data(important_data={'id': updated_obj.id})
 
         self._import_by_csv(csv_data.keys(), [csv_data.values()])
         updated_obj = self.Model.objects.get(pk=updated_obj.id)
@@ -215,8 +217,9 @@ class TestDataImporter(object):
         """
         self.excluded_fields.remove('id')
         updated_obj = self.ModelFactory()
-        csv_data = self._get_csv_data(use_existing=False)
-        csv_data['id'] = updated_obj.id
+        csv_data = self._get_csv_data(
+            important_data={'id': updated_obj.id}, use_existing=False,
+        )
 
         self._import_by_csv(csv_data.keys(), [csv_data.values()])
         updated_obj = self.Model.objects.get(pk=updated_obj.id)
@@ -233,6 +236,33 @@ class TestLicenceDataImporter(TestDataImporter, ClientMixin, TestCase):
         'cache_version', 'created', 'id', 'level', 'lft', 'modified',
         'parent', 'rght', 'saving_user', 'tree_id',
     ])
+
+    def test_add_by_import_with_new_foreign_keys(self):
+        """
+        Creates foreign fields from text values.
+        """
+        self.excluded_fields.update(['region'])
+        csv_data = self._get_csv_data(use_existing=False)
+
+        self._import_by_csv(csv_data.keys(), [csv_data.values()])
+        added_obj = self.Model.objects.latest('id')
+        self._check_object_against_csv(added_obj, csv_data)
+
+    def test_update_by_import_with_new_foreign_fields(self):
+        """
+        Creates foreign fields from text values.
+        """
+        self.excluded_fields.remove('id')
+        self.excluded_fields.update(['region'])
+
+        updated_obj = self.ModelFactory()
+        csv_data = self._get_csv_data(
+            important_data={'id': updated_obj.id}, use_existing=False,
+        )
+
+        self._import_by_csv(csv_data.keys(), [csv_data.values()])
+        updated_obj = self.Model.objects.get(pk=updated_obj.id)
+        self._check_object_against_csv(updated_obj, csv_data)
 
 
 class TestBOAssetDataImporter(TestDataImporter, ClientMixin, TestCase):
@@ -251,29 +281,35 @@ class TestBOAssetDataImporter(TestDataImporter, ClientMixin, TestCase):
         'office_info',
     ])
 
+    def test_add_by_import_with_new_foreign_keys(self):
+        """
+        Creates foreign fields from text values.
+        """
+        self.excluded_fields.update(
+            ['service', 'device_environment', 'region'],
+        )
+        csv_data = self._get_csv_data(use_existing=False)
+
+        self._import_by_csv(csv_data.keys(), [csv_data.values()])
+        added_obj = self.Model.objects.latest('id')
+        self._check_object_against_csv(added_obj, csv_data)
+
     def test_update_by_import_with_new_foreign_fields(self):
         """
         Creates foreign fields from text values.
         """
         self.excluded_fields.remove('id')
-        self.excluded_fields.update(['service', 'device_environment'])
+        self.excluded_fields.update(
+            ['service', 'device_environment', 'region'],
+        )
         updated_obj = self.ModelFactory()
-        csv_data = self._get_csv_data(use_existing=False)
-        csv_data['id'] = updated_obj.id
+        csv_data = self._get_csv_data(
+            important_data={'id': updated_obj.id}, use_existing=False,
+        )
 
         self._import_by_csv(csv_data.keys(), [csv_data.values()])
         updated_obj = self.Model.objects.get(pk=updated_obj.id)
         self._check_object_against_csv(updated_obj, csv_data)
-
-    def test_add_by_import_with_new_foreign_keys(self):
-        """
-        Creates foreign fields from text values.
-        """
-        self.excluded_fields.update(['service', 'device_environment'])
-        csv_data = self._get_csv_data(use_existing=False)
-        self._import_by_csv(csv_data.keys(), [csv_data.values()])
-        added_obj = self.Model.objects.latest('id')
-        self._check_object_against_csv(added_obj, csv_data)
 
 
 class TestDCAssetDataImporter(TestDataImporter, ClientMixin, TestCase):
@@ -299,8 +335,9 @@ class TestDCAssetDataImporter(TestDataImporter, ClientMixin, TestCase):
         Uses existing foreign fields.
         """
         device = DeviceFactory()
-        csv_data = self._get_csv_data()
-        csv_data['barcode'] = device.barcode
+        csv_data = self._get_csv_data(
+            important_data={'barcode': device.barcode},
+        )
         self._import_by_csv(csv_data.keys(), [csv_data.values()])
         added_obj = self.Model.objects.latest('id')
         self._check_object_against_csv(added_obj, csv_data)
@@ -310,13 +347,17 @@ class TestDCAssetDataImporter(TestDataImporter, ClientMixin, TestCase):
         Creates foreign fields from text values.
         """
         device = DeviceFactory()
-        self.excluded_fields.update(['service', 'device_environment'])
-        csv_data = self._get_csv_data(use_existing=False)
-        csv_data['barcode'] = device.barcode
-
         service_env = CIRelationFactory()
-        csv_data['service'] = service_env.parent.name
-        csv_data['device_environment'] = service_env.child.name
+        self.excluded_fields.update(
+            ['service', 'device_environment', 'region'],
+        )
+        csv_data = self._get_csv_data(
+            important_data={
+                'barcode': device.barcode,
+                'device_environment': service_env.child.name,
+                'service': service_env.parent.name,
+            }, use_existing=False,
+        )
 
         self._import_by_csv(csv_data.keys(), [csv_data.values()])
         added_obj = self.Model.objects.latest('id')
@@ -327,10 +368,13 @@ class TestDCAssetDataImporter(TestDataImporter, ClientMixin, TestCase):
         Creates foreign fields from text values.
         """
         self.excluded_fields.remove('id')
-        self.excluded_fields.update(['service', 'device_environment'])
+        self.excluded_fields.update(
+            ['service', 'device_environment', 'region'],
+        )
         updated_obj = self.ModelFactory()
-        csv_data = self._get_csv_data(use_existing=False)
-        csv_data['id'] = updated_obj.id
+        csv_data = self._get_csv_data(
+            important_data={'id': updated_obj.id}, use_existing=False,
+        )
 
         self._import_by_csv(csv_data.keys(), [csv_data.values()])
         updated_obj = self.Model.objects.get(pk=updated_obj.id)