+++ src/ralph_assets/tests/functional/tests_invoice.py
@@ -118,7 +118,7 @@ class TestInvoiceReport(TestCase):
 
     @override_settings(ASSETS_REPORTS=ASSETS_REPORTS_FOR_TESTS)
     @override_settings(INKPY=INKPY_FOR_TESTS)
-    @patch('ralph_assets.views.invoice_report.generate_pdf')
+    @patch('ralph_assets.views.invoice_report.api')
     def test_asset_invoice_with_report_odt_source_fail(self, generate_pdf):
         generate_pdf.side_effect = mocked_generate_pdf
 
@@ -147,7 +147,7 @@ class TestInvoiceReport(TestCase):
 
     @override_settings(ASSETS_REPORTS=ASSETS_REPORTS_FOR_TESTS)
     @override_settings(INKPY=INKPY_FOR_TESTS)
-    @patch('ralph_assets.views.invoice_report.generate_pdf')
+    @patch('ralph_assets.views.invoice_report.api')
     def test_licence_invoice_with_report_odt_source_fail(self, generate_pdf):
         generate_pdf.side_effect = mocked_generate_pdf